Increase number of retries for testing
Tests wait for the webhook_server to come online by polling the
webhook_server until it responds.  10 retries with a 1.5s delay was not
enough for the webhook_server running VirtualBox on my laptop.
Increasing the number of retries fixes the issue for me without slowing
others down in the normal case.