Disambiguate launch port from port used to construct webhooks
Because we use DEIS to deploy, it expects the web process to bind to
port 5000, but it will be available from the outside on port 80.
Therefore we need to construct urls that include port 80, while binding
to port 5000 in the procfile.  We could just hard-code the value in the
procfile, except that we also use this procfile for testing and want to
adjust this port.