Work around race condition in tests
Instead of sleeping, it would be preferable to use a notify transition
to send a webhook telling us that the net should have set the expire on
all keys.  The problem with that approach is that the expire action
MUST be the last action in the net, otherwise more keys will be created
that don't have the expire set on them.  So we sleep, and check again,
if after sleeping we still haven't set the expire then we fail the test.